Assignment1/scripts/requirements.txt
pandas
azure-identity
azure-storage-blob
openpyxl
boto3